Install reanimated
Installed library through expo install & added plugin to babel config as per expo docs. The example taken from the docs did not work at first when running on an existing dev-client. A new client had to be built in order for the example component to work with the newly installed dependencies.