Implement clip duration with scale factor
Clip duration, defined in seconds, will set the trimmer width in such a way that the window is flush with the marker count representing the clip duration.
The scale factor is defined for cases where the clip duration is less than or equal to the 60(track wide marker cap) or greater than 60.